+++ sr_unix/gtm_multi_proc.c
@@ -150,7 +150,7 @@ int	gtm_multi_proc(gtm_multi_proc_fnptr_t fnptr, int ntasks, int max_procs,
 	if (-1 == shmid)
 	{
 		save_errno = errno;
-		SNPRINTF(errstr, SIZEOF(errstr), "shmget() : shmsize=0x%llx", shm_size);
+		SNPRINTF(errstr, SIZEOF(errstr), "shmget() : shmsize=0x%llx", (unsigned long long)shm_size);
 		gtm_putmsg_csa(CSA_ARG(NULL) VARLSTCNT(8) ERR_SYSCALL, 5, LEN_AND_STR(errstr), CALLFROM, save_errno);
 		return -1;
 	}
@@ -158,8 +158,7 @@ int	gtm_multi_proc(gtm_multi_proc_fnptr_t fnptr, int ntasks, int max_procs,
 	if (-1 == (sm_long_t)(multi_proc_shm_hdr))
 	{
 		save_errno = errno;
-		SNPRINTF(errstr, SIZEOF(errstr), "shmat() : shmid=%d shmsize=0x%llx",
-												shmid, shm_size);
+		SNPRINTF(errstr, SIZEOF(errstr), "shmat() : shmid=%d shmsize=0x%llx", shmid, (unsigned long long)shm_size);
 		gtm_putmsg_csa(CSA_ARG(NULL) VARLSTCNT(8) ERR_SYSCALL, 5, LEN_AND_STR(errstr), CALLFROM, save_errno);
 		return -1;
 	}
